move client files
