--- src/client/index.ts
@@ -1,2 +1,0 @@
-export * from "./debug";
-export { store } from "./redux/store";
