+++ src/client/index.tsx
@@ -4,8 +4,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import { App } from "./App";
-import { store } from "./client";
-import { SnackbarProvider } from "./client/snackbar";
+import { store } from "./redux/store";
+import { SnackbarProvider } from "./snackbar";
 
 ReactDOM.render(
   <React.StrictMode>
