+++ src/index.html
@@ -45,7 +45,7 @@
 
   <body>
     <div id="root"></div>
-    <script type="module" src="./src/index.tsx"></script>
+    <script type="module" src="./client/index.tsx"></script>
     <script src="/__/firebase/8.4.3/firebase-app.js"></script>
     <script src="/__/firebase/8.4.3/firebase-analytics.js"></script>
     <script src="/__/firebase/init.js"></script>
